feat: devide login button in to login-button and logged-in-user-button
